requirements.txt
streamlit
pyyaml
pyperclip
